fix: change place for call finctions
