+++ src/components/slider.ts
@@ -7,9 +7,9 @@ const displayValTwo = document.getElementById("range2") as HTMLElement;
 const displayValTree = document.getElementById("range3") as HTMLElement;
 const displayValFour = document.getElementById("range4") as HTMLElement;
 let minGap: number = 0;
-/* let sliderTrack = document.querySelector(".slider-track") as HTMLElement;
+let sliderTrack = document.querySelector(".slider-track") as HTMLElement;
 let sliderTrackTwo = document.querySelector(".slider-track-Two") as HTMLElement;
-let sliderMaxValue : number = parseInt((<HTMLInputElement>document.getElementById("slider-1")).max); */
+let sliderMaxValue : number = parseInt((<HTMLInputElement>document.getElementById("slider-1")).max);
 
 export function slideOne():void{
     if(parseInt(sliderTwo.value) - parseInt(sliderOne.value) <= minGap){
@@ -17,7 +17,7 @@ export function slideOne():void{
         currValueSliderOne = parseInt(sliderTwo.value) - minGap;
     }
     displayValOne.textContent = sliderOne.value;
-    /* fillColor(); */
+    fillColor();
 }
 
 export function slideTwo():void{
@@ -27,7 +27,7 @@ export function slideTwo():void{
         currValueSliderTwo = parseInt(sliderOne.value) + minGap;
     }
     displayValTwo.textContent = sliderTwo.value;
-    /* fillColor(); */
+    fillColor();
 }
 
 export function slideTree():void{
@@ -36,7 +36,7 @@ export function slideTree():void{
         currValueSliderTree = parseInt(sliderFour.value) - minGap;
     }
     displayValTree.textContent = sliderTree.value;
-    /* fillColor2(); */
+    fillColor2();
 }
 
 export function slideFour():void{
@@ -46,15 +46,16 @@ export function slideFour():void{
         currValueSliderTwo = parseInt(sliderTree.value) + minGap;
     }
     displayValFour.textContent = sliderFour.value;
-   /*  fillColor2(); */
+    fillColor2();
 }
 
-/* function fillColor():void{
+function fillColor():void{
   let currValueSliderTwo =  parseInt(sliderTwo.value);
   let currValueSliderOne =  parseInt(sliderOne.value);
   let percent1 = (currValueSliderTwo / sliderMaxValue) * 100;
   let percent2 = (currValueSliderOne / sliderMaxValue) * 100;
-  sliderTrack.style.background = `linear-gradient(to right,  #6a6ad9 ${percent1}% , #6a6ad9 ${percent1}% ,  #6a6ad9 ${percent2}%,  #6a6ad9 ${percent2}%)`;
+  /* sliderTrack.style.background = `linear-gradient(to right,  #6a6ad9 ${percent1}% , #6a6ad9 ${percent1}% ,  #6a6ad9 ${percent2}%,  #6a6ad9 ${percent2}%)`; */
+  console.log(Math.round(percent1 * 1000), Math.round(percent2 * 1000))
 }
 
 function fillColor2():void{
@@ -62,6 +63,7 @@ function fillColor2():void{
     let currValueSliderFour =  parseInt(sliderTree.value);
     let percent1 = (currValueSliderFour / sliderMaxValue) * 100;
     let percent2 = (currValueSliderTree / sliderMaxValue) * 100;
-    sliderTrackTwo.style.background = `linear-gradient(to right, #6a6ad9 ${percent1}% , #3264fe ${percent1}% , #3264fe ${percent2}%, #6a6ad9 ${percent2}%)`;
+    /* sliderTrackTwo.style.background = `linear-gradient(to right, #6a6ad9 ${percent1}% , #3264fe ${percent1}% , #3264fe ${percent2}%, #6a6ad9 ${percent2}%)`; */
+    console.log(Math.round(percent1* 1000),Math.round(percent2*1000))
   }
- */
+
